S2211 (#56)
Remove QueueHandler and use of queue to centralize log messages.

The 'start' and 'end' inputs to _get_obs_file() should be in datetime already.

Fixed issue with Control-C failing intermittently in a multiprocessing environment. This resulted from a bug in Python: https://bugs.python.org/issue8296. The solution was to pass an initializer to the Pool when the Pool was created. The initializer tells each process to ignore SIGINT.